requirements.txt
pandas
numpy
scikit-learn
streamlit
python-binance
pytz
joblib
requests
